+++ handlers/remote_handler.c
@@ -129,7 +129,7 @@ static int RHmsg_get_ack(struct RHmsg *self, void *request)
 	return 0;
 }
 
-static int forward_data(void *request, const void *buf, unsigned int len)
+static int forward_data(void *request, const void *buf, size_t len)
 {
 	struct RHmsg RHmessage;
 	int ret;
