+++ include/util.h
@@ -160,25 +160,25 @@ bool strtobool(const char *s);
 /*
  * Function to extract / copy images
  */
-typedef int (*writeimage) (void *out, const void *buf, unsigned int len);
+typedef int (*writeimage) (void *out, const void *buf, size_t len);
 
 void *saferealloc(void *ptr, size_t size);
 int openfile(const char *filename);
-int copy_write(void *out, const void *buf, unsigned int len);
+int copy_write(void *out, const void *buf, size_t len);
 #if defined(__FreeBSD__)
-int copy_write_padded(void *out, const void *buf, unsigned int len);
+int copy_write_padded(void *out, const void *buf, size_t len);
 #endif
 #if defined(__linux__)
 /* strlcpy was originally developped in FreeBSD, not present in glibc */
 size_t
 strlcpy(char *dst, const char * src, size_t size);
 #endif
-int copyfile(int fdin, void *out, unsigned int nbytes, unsigned long *offs,
+int copyfile(int fdin, void *out, size_t nbytes, unsigned long *offs,
 	unsigned long long seek,
 	int skip_file, int compressed, uint32_t *checksum,
 	unsigned char *hash, bool encrypted, const char *imgivt, writeimage callback);
 int copyimage(void *out, struct img_type *img, writeimage callback);
-int copybuffer(unsigned char *inbuf, void *out, unsigned int nbytes, int compressed,
+int copybuffer(unsigned char *inbuf, void *out, size_t nbytes, int compressed,
 	unsigned char *hash, bool encrypted, const char *imgivt, writeimage callback);
 off_t extract_next_file(int fd, int fdout, off_t start, int compressed,
 			int encrypted, char *ivt, unsigned char *hash);
